+++ libafl_frida/src/asan/asan_rt.rs
@@ -23,12 +23,15 @@ use capstone::{
 #[cfg(target_arch = "x86_64")]
 use capstone::{
     arch::{self, x86::X86OperandType, ArchOperand::X86Operand, BuildsCapstone},
-    Capstone, RegAccessType, RegId,
+    Capstone, RegAccessType, RegId, Insn,
 };
+
+#[cfg(target_arch = "x86_64")]
+use frida_gum::instruction_writer::X86Register;
 
 use dynasmrt::{dynasm, DynasmApi, DynasmLabelApi};
 use frida_gum::interceptor::Interceptor;
-use frida_gum::{Gum, Module, ModuleMap};
+use frida_gum::{Gum, Module, ModuleMap, stalker::StalkerOutput, instruction_writer::InstructionWriter};
 #[cfg(unix)]
 use libc::RLIMIT_STACK;
 use libc::{c_char, wchar_t};
@@ -94,6 +97,7 @@ pub const ASAN_SAVE_REGISTER_COUNT: usize = 32;
 /// even if the target would not have crashed under normal conditions.
 /// this helps finding mem errors early.
 pub struct AsanRuntime {
+    current_report_impl: u64,
     allocator: Allocator,
     regs: [usize; ASAN_SAVE_REGISTER_COUNT],
     blob_report: Option<Box<[u8]>>,
@@ -121,6 +125,7 @@ impl AsanRuntime {
     #[must_use]
     pub fn new(options: FridaOptions) -> AsanRuntime {
         Self {
+            current_report_impl: 0,
             allocator: Allocator::new(options.clone()),
             regs: [0; ASAN_SAVE_REGISTER_COUNT],
             blob_report: None,
@@ -2005,5 +2010,606 @@ impl AsanRuntime {
     #[inline]
     pub fn blob_check_mem_64bytes(&self) -> &[u8] {
         self.blob_check_mem_64bytes.as_ref().unwrap()
+    }
+
+    #[cfg(target_arch = "aarch64")]
+    #[inline]
+    fn writer_register(&self, reg: capstone::RegId) -> Aarch64Register {
+        let regint: u16 = reg.0;
+        Aarch64Register::from_u32(regint as u32).unwrap()
+    }
+
+    // frida registers: https://docs.rs/frida-gum/0.4.0/frida_gum/instruction_writer/enum.X86Register.html
+    // capstone registers: https://docs.rs/capstone-sys/0.14.0/capstone_sys/x86_reg/index.html
+    #[cfg(all(target_arch = "x86_64", unix))]
+    #[must_use]
+    #[inline]
+    #[allow(clippy::unused_self)]
+    pub fn writer_register(&self, reg: RegId) -> X86Register {
+        let regint: u16 = reg.0;
+        match regint {
+            19 => X86Register::Eax,
+            22 => X86Register::Ecx,
+            24 => X86Register::Edx,
+            21 => X86Register::Ebx,
+            30 => X86Register::Esp,
+            20 => X86Register::Ebp,
+            29 => X86Register::Esi,
+            23 => X86Register::Edi,
+            226 => X86Register::R8d,
+            227 => X86Register::R9d,
+            228 => X86Register::R10d,
+            229 => X86Register::R11d,
+            230 => X86Register::R12d,
+            231 => X86Register::R13d,
+            232 => X86Register::R14d,
+            233 => X86Register::R15d,
+            26 => X86Register::Eip,
+            35 => X86Register::Rax,
+            38 => X86Register::Rcx,
+            40 => X86Register::Rdx,
+            37 => X86Register::Rbx,
+            44 => X86Register::Rsp,
+            36 => X86Register::Rbp,
+            43 => X86Register::Rsi,
+            39 => X86Register::Rdi,
+            106 => X86Register::R8,
+            107 => X86Register::R9,
+            108 => X86Register::R10,
+            109 => X86Register::R11,
+            110 => X86Register::R12,
+            111 => X86Register::R13,
+            112 => X86Register::R14,
+            113 => X86Register::R15,
+            41 => X86Register::Rip,
+            _ => X86Register::None, // Ignore Xax..Xip
+        }
+    }
+
+    #[cfg(target_arch = "aarch64")]
+    #[inline]
+    fn instruction_width(&self, instr: &Insn, operands: &Vec<arch::ArchOperand>) -> u32 {
+        use capstone::arch::arm64::Arm64Insn as I;
+        use capstone::arch::arm64::Arm64Reg as R;
+        use capstone::arch::arm64::Arm64Vas as V;
+
+        let num_registers = match instr.id().0.into() {
+            I::ARM64_INS_STP
+            | I::ARM64_INS_STXP
+            | I::ARM64_INS_STNP
+            | I::ARM64_INS_STLXP
+            | I::ARM64_INS_LDP
+            | I::ARM64_INS_LDXP
+            | I::ARM64_INS_LDNP => 2,
+            _ => 1,
+        };
+
+        let mnemonic = instr.mnemonic().unwrap();
+        match mnemonic.as_bytes().last().unwrap() {
+            b'b' => return 1,
+            b'h' => return 2,
+            b'w' => return 4 * num_registers,
+            _ => (),
+        }
+
+        if let Arm64Operand(operand) = operands.first().unwrap() {
+            if operand.vas != V::ARM64_VAS_INVALID {
+                let count_byte: u32 = if mnemonic.starts_with("st") || mnemonic.starts_with("ld") {
+                    mnemonic.chars().nth(2).unwrap().to_digit(10).unwrap()
+                } else {
+                    1
+                };
+
+                return match operand.vas {
+                    V::ARM64_VAS_1B => 1 * count_byte,
+                    V::ARM64_VAS_1H => 2 * count_byte,
+                    V::ARM64_VAS_4B | V::ARM64_VAS_1S | V::ARM64_VAS_1D | V::ARM64_VAS_2H => {
+                        4 * count_byte
+                    }
+                    V::ARM64_VAS_8B
+                    | V::ARM64_VAS_4H
+                    | V::ARM64_VAS_2S
+                    | V::ARM64_VAS_2D
+                    | V::ARM64_VAS_1Q => 8 * count_byte,
+                    V::ARM64_VAS_8H | V::ARM64_VAS_4S | V::ARM64_VAS_16B => 16 * count_byte,
+                    V::ARM64_VAS_INVALID => {
+                        panic!("should not be reached");
+                    }
+                };
+            } else if let Arm64OperandType::Reg(operand) = operand.op_type {
+                match operand.0 as u32 {
+                    R::ARM64_REG_W0..=R::ARM64_REG_W30
+                    | R::ARM64_REG_WZR
+                    | R::ARM64_REG_WSP
+                    | R::ARM64_REG_S0..=R::ARM64_REG_S31 => return 4 * num_registers,
+                    R::ARM64_REG_D0..=R::ARM64_REG_D31 => return 8 * num_registers,
+                    R::ARM64_REG_Q0..=R::ARM64_REG_Q31 => return 16,
+                    _ => (),
+                }
+            };
+        };
+
+        8 * num_registers
+    }
+
+    #[cfg(target_arch = "aarch64")]
+    #[inline]
+    pub fn asan_is_interesting_instruction(
+        &self,
+        capstone: &Capstone,
+        _address: u64,
+        instr: &Insn,
+    ) -> Result<
+        (
+            capstone::RegId,
+            capstone::RegId,
+            i32,
+            u32,
+            Arm64Shift,
+            Arm64Extender,
+        ),
+        (),
+    > {
+        // We have to ignore these instructions. Simulating them with their side effects is
+        // complex, to say the least.
+        match instr.mnemonic().unwrap() {
+            "ldaxr" | "stlxr" | "ldxr" | "stxr" | "ldar" | "stlr" | "ldarb" | "ldarh" | "ldaxp"
+            | "ldaxrb" | "ldaxrh" | "stlrb" | "stlrh" | "stlxp" | "stlxrb" | "stlxrh" | "ldxrb"
+            | "ldxrh" | "stxrb" | "stxrh" => return Err(()),
+            _ => (),
+        }
+
+        let operands = capstone
+            .insn_detail(instr)
+            .unwrap()
+            .arch_detail()
+            .operands();
+        if operands.len() < 2 {
+            return Err(());
+        }
+
+        if let Arm64Operand(arm64operand) = operands.last().unwrap() {
+            if let Arm64OperandType::Mem(opmem) = arm64operand.op_type {
+                return Ok((
+                    opmem.base(),
+                    opmem.index(),
+                    opmem.disp(),
+                    self.instruction_width(instr, &operands),
+                    arm64operand.shift,
+                    arm64operand.ext,
+                ));
+            }
+        }
+
+        Err(())
+    }
+
+    #[cfg(all(target_arch = "x86_64", unix))]
+    #[inline]
+    pub fn asan_is_interesting_instruction(
+        &self,
+        capstone: &Capstone,
+        _address: u64,
+        instr: &Insn,
+    ) -> Result<(RegId, u8, RegId, RegId, i32, i64), ()> {
+        let operands = capstone
+            .insn_detail(instr)
+            .unwrap()
+            .arch_detail()
+            .operands();
+
+        // Ignore lea instruction
+        // put nop into the white-list so that instructions like
+        // like `nop dword [rax + rax]` does not get caught.
+        match instr.mnemonic().unwrap() {
+            "lea" | "nop" => return Err(()),
+
+            _ => (),
+        }
+
+        // This is a TODO! In this case, both the src and the dst are mem operand
+        // so we would need to return two operadns?
+        if instr.mnemonic().unwrap().starts_with("rep") {
+            return Err(());
+        }
+
+        for operand in operands {
+            if let X86Operand(x86operand) = operand {
+                if let X86OperandType::Mem(opmem) = x86operand.op_type {
+                    /*
+                    println!(
+                        "insn: {:#?} {:#?} width: {}, segment: {:#?}, base: {:#?}, index: {:#?}, scale: {}, disp: {}",
+                        insn_id,
+                        instr,
+                        x86operand.size,
+                        opmem.segment(),
+                        opmem.base(),
+                        opmem.index(),
+                        opmem.scale(),
+                        opmem.disp(),
+                    );
+                    */
+                    if opmem.segment() == RegId(0) {
+                        return Ok((
+                            opmem.segment(),
+                            x86operand.size,
+                            opmem.base(),
+                            opmem.index(),
+                            opmem.scale(),
+                            opmem.disp(),
+                        ));
+                    }
+                }
+            }
+        }
+
+        Err(())
+    }
+
+    #[inline]
+    #[allow(clippy::too_many_lines)]
+    #[allow(clippy::too_many_arguments)]
+    #[cfg(all(target_arch = "x86_64", unix))]
+    pub fn emit_shadow_check(
+        &mut self,
+        address: u64,
+        output: &StalkerOutput,
+        _segment: RegId,
+        width: u8,
+        basereg: RegId,
+        indexreg: RegId,
+        scale: i32,
+        disp: i64,
+    ) {
+        let redzone_size = i64::from(frida_gum_sys::GUM_RED_ZONE_SIZE);
+        let writer = output.writer();
+        let true_rip = address;
+
+        let basereg = if basereg.0 == 0 {
+            None
+        } else {
+            let reg = self.writer_register(basereg);
+            Some(reg)
+        };
+
+        let indexreg = if indexreg.0 == 0 {
+            None
+        } else {
+            let reg = self.writer_register(indexreg);
+            Some(reg)
+        };
+
+        let scale = match scale {
+            2 => 1,
+            4 => 2,
+            8 => 3,
+            _ => 0,
+        };
+        if self.current_report_impl == 0
+            || !writer.can_branch_directly_to(self.current_report_impl)
+            || !writer.can_branch_directly_between(writer.pc() + 128, self.current_report_impl)
+        {
+            let after_report_impl = writer.code_offset() + 2;
+
+            #[cfg(target_arch = "x86_64")]
+            writer.put_jmp_near_label(after_report_impl);
+            #[cfg(target_arch = "aarch64")]
+            writer.put_b_label(after_report_impl);
+
+            self.current_report_impl = writer.pc();
+            #[cfg(unix)]
+            writer.put_bytes(self.blob_report());
+
+            writer.put_label(after_report_impl);
+        }
+
+        /* Save registers that we'll use later in shadow_check_blob
+                                        | addr  | rip   |
+                                        | Rcx   | Rax   |
+                                        | Rsi   | Rdx   |
+            Old Rsp - (redsone_size) -> | flags | Rdi   |
+                                        |       |       |
+            Old Rsp                  -> |       |       |
+        */
+        writer.put_lea_reg_reg_offset(X86Register::Rsp, X86Register::Rsp, -(redzone_size));
+        writer.put_pushfx();
+        writer.put_push_reg(X86Register::Rdi);
+        writer.put_push_reg(X86Register::Rsi);
+        writer.put_push_reg(X86Register::Rdx);
+        writer.put_push_reg(X86Register::Rcx);
+        writer.put_push_reg(X86Register::Rax);
+
+        /*
+        Things are a bit different when Rip is either base register or index register.
+        Suppose we have an instruction like
+        `bnd jmp qword ptr [rip + 0x2e4b5]`
+        We can't just emit code like
+        `mov rdi, rip` to get RIP loaded into RDI,
+        because this RIP is NOT the orginal RIP (, which is usually within .text) anymore, rather it is pointing to the memory allocated by the frida stalker.
+        Please confer https://frida.re/docs/stalker/ for details.
+        */
+        // Init Rdi
+        match basereg {
+            Some(reg) => match reg {
+                X86Register::Rip => {
+                    writer.put_mov_reg_address(X86Register::Rdi, true_rip);
+                }
+                _ => {
+                    writer.put_mov_reg_reg(X86Register::Rdi, basereg.unwrap());
+                }
+            },
+            None => {
+                writer.put_xor_reg_reg(X86Register::Rdi, X86Register::Rdi);
+            }
+        }
+
+        match indexreg {
+            Some(reg) => match reg {
+                X86Register::Rip => {
+                    writer.put_mov_reg_address(X86Register::Rsi, true_rip);
+                }
+                _ => {
+                    writer.put_mov_reg_reg(X86Register::Rsi, indexreg.unwrap());
+                }
+            },
+            None => {
+                writer.put_xor_reg_reg(X86Register::Rsi, X86Register::Rsi);
+            }
+        }
+
+        // Scale
+        if scale > 0 {
+            writer.put_shl_reg_u8(X86Register::Rsi, scale);
+        }
+
+        // Finally set Rdi to base + index * scale + disp
+        writer.put_add_reg_reg(X86Register::Rdi, X86Register::Rsi);
+        writer.put_lea_reg_reg_offset(X86Register::Rdi, X86Register::Rdi, disp);
+
+        writer.put_mov_reg_address(X86Register::Rsi, true_rip); // load true_rip into rsi in case we need them in handle_trap
+        writer.put_push_reg(X86Register::Rsi); // save true_rip
+        writer.put_push_reg(X86Register::Rdi); // save accessed_address
+
+        #[cfg(unix)]
+        let checked: bool = match width {
+            1 => writer.put_bytes(self.blob_check_mem_byte()),
+            2 => writer.put_bytes(self.blob_check_mem_halfword()),
+            4 => writer.put_bytes(self.blob_check_mem_dword()),
+            8 => writer.put_bytes(self.blob_check_mem_qword()),
+            16 => writer.put_bytes(self.blob_check_mem_16bytes()),
+            _ => false,
+        };
+
+        if checked {
+            writer.put_jmp_address(self.current_report_impl);
+            for _ in 0..10 {
+                // shadow_check_blob's done will land somewhere in these nops
+                // on amd64 jump can takes 10 bytes at most, so that's why I put 10 bytes.
+                writer.put_nop();
+            }
+        }
+
+        writer.put_pop_reg(X86Register::Rdi);
+        writer.put_pop_reg(X86Register::Rsi);
+
+        writer.put_pop_reg(X86Register::Rax);
+        writer.put_pop_reg(X86Register::Rcx);
+        writer.put_pop_reg(X86Register::Rdx);
+        writer.put_pop_reg(X86Register::Rsi);
+        writer.put_pop_reg(X86Register::Rdi);
+        writer.put_popfx();
+        writer.put_lea_reg_reg_offset(X86Register::Rsp, X86Register::Rsp, redzone_size);
+    }
+
+    #[cfg(target_arch = "aarch64")]
+    #[inline]
+    fn emit_shadow_check(
+        &mut self,
+        _address: u64,
+        output: &StalkerOutput,
+        basereg: capstone::RegId,
+        indexreg: capstone::RegId,
+        displacement: i32,
+        width: u32,
+        shift: Arm64Shift,
+        extender: Arm64Extender,
+    ) {
+        let redzone_size = frida_gum_sys::GUM_RED_ZONE_SIZE as i32;
+        let writer = output.writer();
+
+        let basereg = self.writer_register(basereg);
+        let indexreg = if indexreg.0 != 0 {
+            Some(self.writer_register(indexreg))
+        } else {
+            None
+        };
+
+        if self.current_report_impl == 0
+            || !writer.can_branch_directly_to(self.current_report_impl)
+            || !writer.can_branch_directly_between(writer.pc() + 128, self.current_report_impl)
+        {
+            let after_report_impl = writer.code_offset() + 2;
+
+            #[cfg(target_arch = "x86_64")]
+            writer.put_jmp_near_label(after_report_impl);
+            #[cfg(target_arch = "aarch64")]
+            writer.put_b_label(after_report_impl);
+
+            self.current_report_impl = writer.pc();
+
+            #[cfg(unix)]
+            writer.put_bytes(self.asan_runtime.blob_report());
+
+            writer.put_label(after_report_impl);
+        }
+        //writer.put_brk_imm(1);
+
+        // Preserve x0, x1:
+        writer.put_stp_reg_reg_reg_offset(
+            Aarch64Register::X0,
+            Aarch64Register::X1,
+            Aarch64Register::Sp,
+            -(16 + redzone_size) as i64,
+            IndexMode::PreAdjust,
+        );
+
+        // Make sure the base register is copied into x0
+        match basereg {
+            Aarch64Register::X0 | Aarch64Register::W0 => {}
+            Aarch64Register::X1 | Aarch64Register::W1 => {
+                writer.put_mov_reg_reg(Aarch64Register::X0, Aarch64Register::X1);
+            }
+            _ => {
+                if !writer.put_mov_reg_reg(Aarch64Register::X0, basereg) {
+                    writer.put_mov_reg_reg(Aarch64Register::W0, basereg);
+                }
+            }
+        }
+
+        // Make sure the index register is copied into x1
+        if indexreg.is_some() {
+            if let Some(indexreg) = indexreg {
+                match indexreg {
+                    Aarch64Register::X0 | Aarch64Register::W0 => {
+                        writer.put_ldr_reg_reg_offset(
+                            Aarch64Register::X1,
+                            Aarch64Register::Sp,
+                            0u64,
+                        );
+                    }
+                    Aarch64Register::X1 | Aarch64Register::W1 => {}
+                    _ => {
+                        if !writer.put_mov_reg_reg(Aarch64Register::X1, indexreg) {
+                            writer.put_mov_reg_reg(Aarch64Register::W1, indexreg);
+                        }
+                    }
+                }
+            }
+
+            if let (Arm64Extender::ARM64_EXT_INVALID, Arm64Shift::Invalid) = (extender, shift) {
+                writer.put_add_reg_reg_reg(
+                    Aarch64Register::X0,
+                    Aarch64Register::X0,
+                    Aarch64Register::X1,
+                );
+            } else {
+                let extender_encoding: i32 = match extender {
+                    Arm64Extender::ARM64_EXT_UXTB => 0b000,
+                    Arm64Extender::ARM64_EXT_UXTH => 0b001,
+                    Arm64Extender::ARM64_EXT_UXTW => 0b010,
+                    Arm64Extender::ARM64_EXT_UXTX => 0b011,
+                    Arm64Extender::ARM64_EXT_SXTB => 0b100,
+                    Arm64Extender::ARM64_EXT_SXTH => 0b101,
+                    Arm64Extender::ARM64_EXT_SXTW => 0b110,
+                    Arm64Extender::ARM64_EXT_SXTX => 0b111,
+                    _ => -1,
+                };
+                let (shift_encoding, shift_amount): (i32, u32) = match shift {
+                    Arm64Shift::Lsl(amount) => (0b00, amount),
+                    Arm64Shift::Lsr(amount) => (0b01, amount),
+                    Arm64Shift::Asr(amount) => (0b10, amount),
+                    _ => (-1, 0),
+                };
+
+                if extender_encoding != -1 && shift_amount < 0b1000 {
+                    // emit add extended register: https://developer.arm.com/documentation/ddi0602/latest/Base-Instructions/ADD--extended-register---Add--extended-register--
+                    writer.put_bytes(
+                        &(0x8b210000 | ((extender_encoding as u32) << 13) | (shift_amount << 10))
+                            .to_le_bytes(),
+                    );
+                } else if shift_encoding != -1 {
+                    writer.put_bytes(
+                        &(0x8b010000 | ((shift_encoding as u32) << 22) | (shift_amount << 10))
+                            .to_le_bytes(),
+                    );
+                } else {
+                    panic!("extender: {:?}, shift: {:?}", extender, shift);
+                }
+            };
+        }
+
+        let displacement = displacement
+            + if basereg == Aarch64Register::Sp {
+                16 + redzone_size
+            } else {
+                0
+            };
+
+        #[allow(clippy::comparison_chain)]
+        if displacement < 0 {
+            if displacement > -4096 {
+                // Subtract the displacement into x0
+                writer.put_sub_reg_reg_imm(
+                    Aarch64Register::X0,
+                    Aarch64Register::X0,
+                    displacement.abs() as u64,
+                );
+            } else {
+                let displacement_hi = displacement.abs() / 4096;
+                let displacement_lo = displacement.abs() % 4096;
+                writer.put_bytes(&(0xd1400000u32 | ((displacement_hi as u32) << 10)).to_le_bytes());
+                writer.put_sub_reg_reg_imm(
+                    Aarch64Register::X0,
+                    Aarch64Register::X0,
+                    displacement_lo as u64,
+                );
+            }
+        } else if displacement > 0 {
+            if displacement < 4096 {
+                // Add the displacement into x0
+                writer.put_add_reg_reg_imm(
+                    Aarch64Register::X0,
+                    Aarch64Register::X0,
+                    displacement as u64,
+                );
+            } else {
+                let displacement_hi = displacement / 4096;
+                let displacement_lo = displacement % 4096;
+                writer.put_bytes(&(0x91400000u32 | ((displacement_hi as u32) << 10)).to_le_bytes());
+                writer.put_add_reg_reg_imm(
+                    Aarch64Register::X0,
+                    Aarch64Register::X0,
+                    displacement_lo as u64,
+                );
+            }
+        }
+        // Insert the check_shadow_mem code blob
+        #[cfg(unix)]
+        match width {
+            1 => writer.put_bytes(&self.asan_runtime.blob_check_mem_byte()),
+            2 => writer.put_bytes(&self.asan_runtime.blob_check_mem_halfword()),
+            3 => writer.put_bytes(&self.asan_runtime.blob_check_mem_3bytes()),
+            4 => writer.put_bytes(&self.asan_runtime.blob_check_mem_dword()),
+            6 => writer.put_bytes(&self.asan_runtime.blob_check_mem_6bytes()),
+            8 => writer.put_bytes(&self.asan_runtime.blob_check_mem_qword()),
+            12 => writer.put_bytes(&self.asan_runtime.blob_check_mem_12bytes()),
+            16 => writer.put_bytes(&self.asan_runtime.blob_check_mem_16bytes()),
+            24 => writer.put_bytes(&self.asan_runtime.blob_check_mem_24bytes()),
+            32 => writer.put_bytes(&self.asan_runtime.blob_check_mem_32bytes()),
+            48 => writer.put_bytes(&self.asan_runtime.blob_check_mem_48bytes()),
+            64 => writer.put_bytes(&self.asan_runtime.blob_check_mem_64bytes()),
+            _ => false,
+        };
+
+        // Add the branch to report
+        //writer.put_brk_imm(0x12);
+        writer.put_branch_address(self.current_report_impl);
+
+        match width {
+            3 | 6 | 12 | 24 | 32 | 48 | 64 => {
+                let msr_nvcz_x0: u32 = 0xd51b4200;
+                writer.put_bytes(&msr_nvcz_x0.to_le_bytes());
+            }
+            _ => (),
+        }
+
+        // Restore x0, x1
+        assert!(writer.put_ldp_reg_reg_reg_offset(
+            Aarch64Register::X0,
+            Aarch64Register::X1,
+            Aarch64Register::Sp,
+            16 + redzone_size as i64,
+            IndexMode::PostAdjust,
+        ));
     }
 }
